+++ components/src/core/DrawerLayout/contexts/DrawerLayoutContextProvider.tsx
@@ -1,13 +1,13 @@
 import { createContext, useContext } from 'react';
 
 type DrawerLayoutContextType = {
-    setPadding: (padding: number) => void;
+    setPadding: (padding: number | string) => void;
     setDrawerOpen: (open: boolean) => void;
 };
 
 export const DrawerLayoutContext = createContext<DrawerLayoutContextType | null>({
     /* eslint-disable @typescript-eslint/no-unused-vars */
-    setPadding: (padding: number) => null,
+    setPadding: (padding: number | string) => null,
     setDrawerOpen: (open: boolean) => false,
     /* eslint-enable @typescript-eslint/no-unused-vars */
 });