+++ components/src/core/Hero/Hero.test.tsx
@@ -5,7 +5,8 @@ import { Mount, Shallow } from '../types';
 import { Hero } from './Hero';
 import { ChannelValue } from '../ChannelValue';
 import Enzyme from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
+// import Adapter from 'enzyme-adapter-react-16';
+import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
 import { createMount, createShallow } from '@material-ui/core/test-utils';
 
 Enzyme.configure({ adapter: new Adapter() });