+++ components/src/core/ListItemTag/ListItemTag.tsx
@@ -30,13 +30,20 @@ declare module '@material-ui/core/styles/withStyles' {
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
         root: {
-            borderRadius: theme.spacing(0.25),
+            borderRadius: '0.125rem',
             padding: 0,
             paddingLeft: theme.spacing(0.5),
             paddingRight: theme.spacing(0.5) - 1, // to account for extra pixel from letter-spacing
             overflow: 'hidden',
-            backgroundColor: (props: ListItemTagProps): string => props.backgroundColor || theme.palette.primary.main,
-            color: (props: ListItemTagProps): string => props.fontColor || theme.palette.primary.contrastText,
+            backgroundColor: (props: ListItemTagProps): string =>
+                props.backgroundColor ||
+                (theme.palette.type === 'dark' ? theme.palette.primary.dark : theme.palette.primary.main),
+            color: (props: ListItemTagProps): string =>
+                props.fontColor ||
+                theme.palette.getContrastText(
+                    props.backgroundColor ||
+                        (theme.palette.type === 'dark' ? theme.palette.primary.dark : theme.palette.primary.main)
+                ),
             cursor: (props: ListItemTagProps): string => (props.onClick ? 'pointer' : 'inherit'),
             display: 'inline-block',
             flip: false, // letter-spacing doesn't flip for RTL, so neither shall our padding hack to offset it
@@ -44,9 +51,9 @@ const useStyles = makeStyles((theme: Theme) =>
         noVariant: {
             fontWeight: 700, // bold
             letterSpacing: 1,
-            fontSize: 10,
-            lineHeight: '16px',
-            height: 16,
+            fontSize: `0.625rem`,
+            lineHeight: `1rem`,
+            height: `1rem`,
         },
     })
 );