+++ components/src/core/UserMenu/UserMenu.tsx
@@ -1,36 +1,37 @@
 import { Menu, MenuProps as standardMenuProps, useTheme } from '@material-ui/core';
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+import { createStyles, makeStyles } from '@material-ui/core/styles';
 import clsx from 'clsx';
 import React, { useCallback, useState, useEffect, HTMLAttributes } from 'react';
 import PropTypes from 'prop-types';
 import { DrawerHeader, DrawerNavGroup, NavItem } from '../Drawer';
 
-const useStyles = makeStyles((theme: Theme) =>
+const useStyles = makeStyles(() =>
     createStyles({
         root: {},
-        navGroups: {
-            '&:active, &:focus': {
-                outline: 'none',
-            },
+        avatarRoot: {
+            cursor: 'pointer',
+            height: `2.5rem`,
+            width: `2.5rem`,
         },
         header: {
             '&:active, &:focus': {
                 outline: 'none',
             },
         },
-        avatarRoot: {
-            cursor: 'pointer',
-            backgroundColor: theme.palette.primary.light,
-            color: theme.palette.primary.main,
-            height: theme.spacing(5),
-            width: theme.spacing(5),
+        menuTitle: {
+            fontSize: '1rem',
+            lineHeight: 1.4,
+        },
+        navigation: {
+            alignSelf: 'center',
+        },
+        navGroups: {
+            '&:active, &:focus': {
+                outline: 'none',
+            },
         },
         noCursor: {
             cursor: 'inherit',
-        },
-        margins: {
-            marginLeft: -theme.spacing(1),
-            marginRight: -theme.spacing(1),
         },
     })
 );
@@ -148,8 +149,10 @@ const UserMenuRender: React.ForwardRefRenderFunction<unknown, UserMenuProps> = (
                         subtitle={menuSubtitle}
                         fontColor={'inherit'}
                         backgroundColor={'inherit'}
+                        divider
                         classes={{
-                            nonClickableIcon: defaultClasses.margins,
+                            title: defaultClasses.menuTitle,
+                            navigation: defaultClasses.navigation,
                         }}
                     />
                 </div>
@@ -169,7 +172,12 @@ const UserMenuRender: React.ForwardRefRenderFunction<unknown, UserMenuProps> = (
                         title={group.title}
                         items={group.items.map(
                             (item: UserMenuItem, itemIndex: number): NavItem =>
-                                Object.assign({ itemID: itemIndex.toString() }, item)
+                                Object.assign({ itemID: itemIndex.toString() }, item, {
+                                    InfoListItemProps: Object.assign(
+                                        { iconColor: theme.palette.text.secondary },
+                                        item.InfoListItemProps
+                                    ),
+                                })
                         )}
                     />
                 </div>