+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 5.1.0
+
+### Changed
+
+-   Some default colors in `<InfoListItem>`, `<Drawer>` and `<ScoreCard>` headers to work with @pxblue/react-themes version 6+.
+-   Some default styles on `<ScoreCard>`, `<UserMenu>`, and `<EmptyState>` text elements.
+-   Component sizing updated to use rems to better respond to system font size changes.
+
+### Added
+
+-   `openOnHoverDelay` to `<Drawer>` to alter open-on-hover delay for closed persistent drawers. 
+-   peerDependency support for using with React 17
+
 ## 5.0.0
 
 ### Changed
@@ -11,6 +24,7 @@
 -   Some of the class names for style overrides have changed
 
 ### Added
+
 -   Additional configuration properties for `<Drawer>`:
     -   `sideBorder` gives the `<Drawer>` a slight border instead of a drop shadow.
     -   `disableActiveItemParentStyle` disables the bold text style for active item's parent elements.
@@ -59,10 +73,12 @@
 
 ## v4.0.2
 
-### Added 
+### Added
+
 -   Adds `info` prop to `<InfoListItem>` to support a third line of text.
 
 ### Changed
+
 -   Updates several prop types to `ReactNode` to support wider range of input values.
 
 ## v4.0.1
@@ -71,10 +87,12 @@
 
 ## v4.0.0
 
-### Added  
+### Added
+
 -   Style-related properties are now overridable by the `classes` prop in each component.
 
 ### Changed
+
 -   Supplemental properties are now spread to the root component for each PX Blue component.
 -   Changed component font colors to address potential accessibility issues.
 
@@ -85,16 +103,19 @@
 ## v3.0.3
 
 ### Fixed
+
 -   Fixes IE 11 issue where persistent `<Drawer>` will not close.
 -   Fixes some spacing issues when using the `<DrawerLayout>`.
 
 ## v3.0.0
 
-### Added 
+### Added
+
 -   Adds support for nested items in the `<Drawer>` component.
 -   Additional styling props added to `<Drawer>`.
 
-### Changed 
+### Changed
+
 **Breaking Changes:**
 
 -   A few props got renamed to avoid further ambiguities:
@@ -137,18 +158,21 @@
 
 ## v2.1.0
 
-### Added 
+### Added
+
 -   Adds `<InfoListTag>` Component
     -   Displays additional information inside an InfoListItem.
 -   Adds `<UserMenu>` Component
     -   Avatar which opens a Menu when clicked.
-    
-### Fixed 
+
+### Fixed
+
 -   Misc bug fixes.
 
 ## v2.0.0
 
 ### Added
+
 -   Adds new components for:
     -   `<Drawer>`
     -   `<DrawerHeader>`
@@ -159,6 +183,7 @@
     -   `<DrawerLayout>`
 
 ### Changed
+
 -   Library converted to TypeScript to provide strong typings for TS projects.
 -   **Breaking Change:** Simpler import syntax - _default_ imports will no longer work.
 
@@ -172,12 +197,14 @@
 
 ## v1.1.0
 
-### Added 
+### Added
+
 Adds a new component for `<EmptyState>`.
 
 ## v1.0.0
 
-### Fixed 
+### Fixed
+
 Fixes a bug in icon size for inline `<ChannelValue>` components.
 
 ## v0.0.1